maint: Move tests and fixtures
