+++ .gitmodules
@@ -1,3 +1,3 @@
-[submodule "tests/wycheproof"]
-	path = tests/wycheproof
+[submodule "tests/integration/WycheProof/fixtures"]
+	path = tests/integration/WycheProof/fixtures
 	url = https://github.com/google/wycheproof
