+++ composer.json
@@ -46,8 +46,8 @@
     },
     "autoload-dev": {
         "psr-4": {
-            "Mdanter\\Ecc\\Tests\\": "tests/unit",
-            "Mdanter\\Ecc\\WycheProof\\": "tests/wycheproof"
+            "Mdanter\\Ecc\\Tests": "tests/unit",
+            "Mdanter\\Ecc\\Integration\\": "tests/integration"
         }
     }
 }
