+++ tests/integration/Spec/fixtures/nist-p192.yml
