+++ tests/integration/Spec/fixtures/secp-192k1.yml
