+++ tests/integration/WycheProof/AbstractTestCase.php
@@ -1,7 +1,7 @@
 <?php
 declare(strict_types=1);
 
-namespace Mdanter\Ecc\WycheProof;
+namespace Mdanter\Ecc\Integration\WycheProof;
 
 use Mdanter\Ecc\Crypto\Signature\HasherInterface;
 use Mdanter\Ecc\Crypto\Signature\SignHasher;
