+++ tests/integration/WycheProof/EcdhTest.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace Mdanter\Ecc\WycheProof;
+namespace Mdanter\Ecc\Integration\WycheProof;
 
 use Mdanter\Ecc\Curves\CurveFactory;
 use Mdanter\Ecc\Exception\ExchangeException;
@@ -38,7 +38,7 @@ class EcdhTest extends AbstractTestCase
 
     public function getEcDHFixtures(): array
     {
-        $fixtures = json_decode($this->importFile("import/wycheproof/testvectors/ecdh_test.json"), true);
+        $fixtures = json_decode($this->importFile("wycheproof/wycheproof/testvectors/ecdh_test.json"), true);
         return $this->filterFixtures($fixtures, $this->getCurvesList());
     }
 
@@ -115,7 +115,7 @@ class EcdhTest extends AbstractTestCase
 
     public function getSpecificFixtures(string $curve): array
     {
-        $fixtures = json_decode($this->importFile("import/wycheproof/testvectors/ecdh_{$curve}_test.json"), true);
+        $fixtures = json_decode($this->importFile("wycheproof/wycheproof/testvectors/ecdh_{$curve}_test.json"), true);
         $filtered =  $this->filterFixtures($fixtures);
 
         return $filtered;
